Rename filesystem to db_pathadn refactor code a bit
